Use AWS CLI for bootstrap target
The (former) boostrap target was using terraform which forced
the user to put a terraform state file under version control.
Switch to using AWS CLI instead since it is just a matter of
creating a bucket and a dynamodb table. Also put a README file
in both state buckets created as a help to the end user.

Rename the target to prereqs not to collide with the project
name.